chore(dal): move RootProp to its own module
- Move RootProp to its own module under SchemaVariant
  - It previously existed under "built-ins", but was used in multiple
    areas of the codebase
- Add module doc comments to EditField, Schema "built-ins", and the new
  RootProp module
- Add doc comments to Component functions related to EditFields
  - We need to indicate to developers that EditFields are data bags that
    leverage AttributeValues for meaningful operations
